OriginalProjectBackup/db-proxy-client/requirements.txt
websockets
mysql-connector-python
python-dotenv
